Extend the former ConstDebuginfo to handle locals that are only assigned once.
This should reduce dependencies between debuginfo and locals, letting other passes such as `SimplifyLocals` do more work.

Fixes #78994
