+++ compiler/rustc_mir/src/transform/mod.rs
@@ -21,7 +21,6 @@ pub mod check_consts;
 pub mod check_packed_ref;
 pub mod check_unsafety;
 pub mod cleanup_post_borrowck;
-pub mod const_debuginfo;
 pub mod const_goto;
 pub mod const_prop;
 pub mod coverage;
@@ -50,6 +49,7 @@ pub mod rustc_peek;
 pub mod simplify;
 pub mod simplify_branches;
 pub mod simplify_comparison_integral;
+pub mod simplify_debuginfo;
 pub mod simplify_try;
 pub mod uninhabited_enum_branching;
 pub mod unreachable_prop;
@@ -510,7 +510,7 @@ fn run_optimization_passes<'tcx>(tcx: TyCtxt<'tcx>, body: &mut Body<'tcx>) {
         &remove_noop_landing_pads::RemoveNoopLandingPads,
         &simplify::SimplifyCfg::new("final"),
         &nrvo::RenameReturnPlace,
-        &const_debuginfo::ConstDebugInfo,
+        &simplify_debuginfo::SimplifyDebugInfo,
         &simplify::SimplifyLocals,
         &multiple_return_terminators::MultipleReturnTerminators,
         &deduplicate_blocks::DeduplicateBlocks,
