+++ src/test/mir-opt/inline/inline_closure_borrows_arg.foo.Inline.after.mir
@@ -18,7 +18,7 @@ fn foo(_1: T, _2: &i32) -> i32 {
             debug _s => _9;              // in scope 2 at $DIR/inline-closure-borrows-arg.rs:16:5: 16:12
             let _10: &i32;               // in scope 2 at $DIR/inline-closure-borrows-arg.rs:16:5: 16:12
             scope 3 {
-                debug variable => _10;   // in scope 3 at $DIR/inline-closure-borrows-arg.rs:16:5: 16:12
+                debug variable => _8;    // in scope 3 at $DIR/inline-closure-borrows-arg.rs:16:5: 16:12
             }
         }
     }
