+++ src/test/mir-opt/simplify_debuginfo_moves.rs
@@ -1,0 +1,6 @@
+#![crate_type = "lib"]
+
+// EMIT_MIR simplify_debuginfo_moves.test_other.SimplifyDebugInfo.diff
+pub fn test_other(x: &Vec<usize>) -> usize {
+    x.iter().collect::<Vec<_>>().len()
+}
