+++ src/test/mir-opt/simplify_try.try_identity.SimplifyLocals.after.mir
@@ -19,7 +19,7 @@ fn try_identity(_1: Result<u32, i32>) -> Result<u32, i32> {
         debug v => ((_0 as Ok).0: u32);  // in scope 3 at $DIR/simplify_try.rs:23:12: 23:13
     }
     scope 4 (inlined into_result::<u32, i32>) { // at $DIR/simplify_try.rs:21:19: 21:33
-        debug r => _0;                   // in scope 4 at $DIR/simplify_try.rs:21:19: 21:33
+        debug r => _1;                   // in scope 4 at $DIR/simplify_try.rs:21:19: 21:33
     }
 
     bb0: {
